Refactor: #160 アカウントの`other_settings`関連部分を別ファイルに切り出し
